# Random patch sampling